Add missing file extensions and make folder names more consistent
