+++ C++/Prims-algo.cpp
