+++ C/Rhombus_patten.c
